컨벤션 내용 적용  - ServiceImpl 삭제  - dto/RefreshToken -> entity 디렉토리 하위로 이동  - type 디렉토리 common 디렉토리 하위로 이동
