--- .github/workflows/development.yml
@@ -1,97 +1,0 @@
-name: CI/CD with Gradle
-
-on:
-  push:
-    branches: [ "develop" ]
-  pull_request:
-    branches: [ "develop" ]
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up JDK 11
-        uses: actions/setup-java@v2
-        with:
-          java-version: '11'
-          distribution: 'temurin'
-
-      - name: Grant execute permission for gradlew
-        run: chmod +x gradlew
-
-      - name: Run build
-        run: ./gradlew clean build -x test
-
-      - name: Login to DockerHub
-        uses: docker/login-action@v1
-        with:
-          username: ${{secrets.DOCKERHUB_USERNAME}}
-          password: ${{secrets.DOCKERHUB_TOKEN}}
-
-      - name: Docker build & push
-        run: |
-          docker build -t ${{secrets.DOCKER_REPO}} .
-          docker tag ${{secrets.DOCKER_REPO}}:latest ${{secrets.DOCKERHUB_USERNAME}}/${{secrets.DOCKER_REPO}}:latest
-          docker push ${{secrets.DOCKERHUB_USERNAME}}/${{secrets.DOCKER_REPO}}
-
-      # notification slack
-      - uses: ravsamhq/notify-slack-action@v1
-        if: ${{ failure() }}
-        with:
-          notification_title: '*execute ${{ github.job }} by ${{ github.actor }}*'
-          notify_when: 'failure,warnings'
-          icon_success: ':white_check_mark:'
-          message_format: |-
-            *workflow*: {workflow}
-            *status*: {emoji} {status_message} 
-            *branch*: <{repo_url}|{branch}>
-            *commit*: <{commit_url}|{commit_sha}>
-            *commit message*: ${{ github.event.head_commit.message }}
-          status: ${{ job.status }}
-          mention_users: 'U03L1TSLD2P,U03KTKP91KP'
-          mention_users_when: 'failure'
-          footer: '<{run_url}|View Github Actions>'
-        env:
-          SLACK_WEBHOOK_URL: ${{ secrets.SLACK_WEBHOOK_URL }}
-
-  deploy:
-    needs: build
-    name: Deploy
-    runs-on: [ self-hosted ]
-    steps:
-      - name: Login to DockerHub
-        uses: docker/login-action@v1
-        with:
-          username: ${{secrets.DOCKERHUB_USERNAME}}
-          password: ${{secrets.DOCKERHUB_TOKEN}}
-      - name: Run container
-        run: |
-          CONTAINER_LIST=$(docker ps -q -a)
-          if [ -z "$CONTAINER_LIST" ];
-          then
-            echo "docker container not running."
-          else
-            docker rm -f $(docker ps -q -a)
-          fi
-          docker pull ${{secrets.DOCKERHUB_USERNAME}}/${{secrets.DOCKER_REPO}}
-          docker run -itd -p 80:8080 --name 11-02-backend ${{secrets.DOCKERHUB_USERNAME}}/${{secrets.DOCKER_REPO}}:latest
-
-      # notification slack
-      - uses: ravsamhq/notify-slack-action@v1
-        if: ${{ failure() }}
-        with:
-          notification_title: '*execute ${{ github.job }} by ${{ github.actor }}*'
-          icon_success: ':white_check_mark:'
-          message_format: |-
-            *workflow*: {workflow}
-            *status*: {emoji} {status_message} 
-            *branch*: <{repo_url}|{branch}>
-            *commit*: <{commit_url}|{commit_sha}>
-            *commit message*: ${{ github.event.head_commit.message }}
-          status: ${{ job.status }}
-          mention_users: 'U03L1TSLD2P,U03KTKP91KP'
-          mention_users_when: 'failure'
-          footer: '<{run_url}|View Github Actions>'
-        env:
-          SLACK_WEBHOOK_URL: ${{ secrets.SLACK_WEBHOOK_URL }}
