--- Dockerfile
@@ -1,6 +1,0 @@
-FROM openjdk:11
-
-COPY build/libs/bside-0.0.1-SNAPSHOT.jar app.jar
-
-ENTRYPOINT ["java", "-jar", "app.jar"]
-
