+++ src/main/kotlin/com/bside/service/MemberService.kt
@@ -1,13 +1,23 @@
 package com.bside.service
 
 import com.bside.dto.MemberResponseDto
+import com.bside.repository.MemberReposiroty
+import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
+
 
 /**
- * name : MemberService
+ * name : MemberServiceImpl
  * author : jisun.noh
  */
-interface MemberService {
-
-    fun getMemberInfo(email: String): MemberResponseDto
-
+@Service
+@Transactional
+class MemberService(
+    val userRepository: MemberReposiroty
+) {
+    fun getMemberInfo(email: String): MemberResponseDto {
+        return MemberResponseDto().apply {
+            this.email = userRepository.findByEmail(email).email
+        }
+    }
 }
